Split parsers file
